[c2goto] library: rename *.{hs => h}, add __clang__ version of stddef.h, move pthreadtypes.h to bits/
