+++ src/c2goto/headers/CMakeLists.txt
@@ -1,4 +1,4 @@
-set (inputs limits.hs pthread.hs pthreadtypes.hs stdarg.hs stdbool.hs stddef.hs)
+set (inputs limits.h pthread.h bits/pthreadtypes.h stdarg.h stdbool.h stddef.h)
 mangle(mangled_files ${CMAKE_CURRENT_SOURCE_DIR} ${inputs}
        PREFIX esbmc_libc_header_
        MACRO ESBMC_FLAIL
